ingest/requirements.txt
requests
python-dotenv
peewee
psycopg2-binary
